Replace %zu in log format

+++ src/app/InteractionModelEngine.cpp
@@ -1588,16 +1588,16 @@ CHIP_ERROR InteractionModelEngine::ResumeSubscriptions()
 #if CHIP_CONFIG_PERSIST_SUBSCRIPTIONS
     SubscriptionResumptionStorage::SubscriptionIndex subscriberIndex;
     CHIP_ERROR err = mpSubscriptionResumptionStorage->LoadIndex(subscriberIndex);
-    ChipLogProgress(InteractionModel, "%zu subscriber nodes to resume.. (error %" CHIP_ERROR_FORMAT ")", subscriberIndex.mSize,
-                    err.Format());
+    ChipLogProgress(InteractionModel, "%u subscriber nodes to resume.. (error %" CHIP_ERROR_FORMAT ")",
+                    static_cast<unsigned>(subscriberIndex.mSize), err.Format());
     for (size_t i = 0; i < subscriberIndex.mSize; i++)
     {
         std::vector<SubscriptionResumptionStorage::SubscriptionInfo> subscriptions;
         err = mpSubscriptionResumptionStorage->FindByScopedNodeId(subscriberIndex.mNodes[i], subscriptions);
 
-        ChipLogProgress(InteractionModel,
-                        "\tNode " ChipLogFormatScopedNodeId ": Loaded %zu subscriptions.. (error %" CHIP_ERROR_FORMAT ")",
-                        ChipLogValueScopedNodeId(subscriberIndex.mNodes[i]), subscriptions.size(), err.Format());
+        ChipLogProgress(
+            InteractionModel, "\tNode " ChipLogFormatScopedNodeId ": Loaded %u subscriptions.. (error %" CHIP_ERROR_FORMAT ")",
+            ChipLogValueScopedNodeId(subscriberIndex.mNodes[i]), static_cast<unsigned>(subscriptions.size()), err.Format());
 
         for (auto & subscriptionInfo : subscriptions)
         {
